backend/requirements.txt
django
djangorestframework
djangorestframework-jwt
django-rest-auth
django-cors-headers
markdown
django-filter
pudb
